+++ packages/react-avatar/src/index.ts
@@ -9,3 +9,11 @@ export {
 } from './Avatar';
 export type { AvatarNamedColor, AvatarProps, AvatarSlots, AvatarState } from './Avatar';
 export { getInitials } from './utils/index';
+export {
+  AvatarGroup,
+  avatarGroupClassNames,
+  renderAvatarGroup_unstable,
+  useAvatarGroupStyles_unstable,
+  useAvatarGroup_unstable,
+} from './AvatarGroup';
+export type { AvatarGroupProps, AvatarGroupSlots, AvatarGroupState } from './AvatarGroup';