+++ packages/react-combobox/src/components/Combobox/Combobox.test.tsx
@@ -7,6 +7,20 @@ describe('Combobox', () => {
   isConformant({
     Component: Combobox,
     displayName: 'Combobox',
+    primarySlot: 'trigger',
+    // don't test deprecated className export on new components
+    disabledTests: ['component-has-static-classname-exported'],
+    testOptions: {
+      'has-static-classnames': [
+        {
+          props: {
+            open: true,
+            // Portal messes with the classNames test, so rendering the listbox inline here
+            inline: true,
+          },
+        },
+      ],
+    },
   });
 
   // TODO add more tests here, and create visual regression tests in /apps/vr-tests