+++ packages/react-combobox/src/components/Listbox/Listbox.test.tsx
@@ -7,6 +7,8 @@ describe('Listbox', () => {
   isConformant({
     Component: Listbox,
     displayName: 'Listbox',
+    // don't test deprecated className export on new components
+    disabledTests: ['component-has-static-classname-exported'],
   });
 
   // TODO add more tests here, and create visual regression tests in /apps/vr-tests