+++ packages/react-combobox/src/components/OptionGroup/OptionGroup.test.tsx
@@ -7,6 +7,17 @@ describe('OptionGroup', () => {
   isConformant({
     Component: OptionGroup,
     displayName: 'OptionGroup',
+    // don't test deprecated className export on new components
+    disabledTests: ['component-has-static-classname-exported'],
+    testOptions: {
+      'has-static-classnames': [
+        {
+          props: {
+            label: 'group label',
+          },
+        },
+      ],
+    },
   });
 
   // TODO add more tests here, and create visual regression tests in /apps/vr-tests