+++ packages/react-menu/config/tests.js
@@ -1,9 +1,3 @@
 /** Jest test setup file. */
 
 require('@testing-library/jest-dom');
-
-const { configure } = require('enzyme');
-const Adapter = require('enzyme-adapter-react-16');
-
-// Configure enzyme.
-configure({ adapter: new Adapter() });