+++ packages/react-toolbar/src/components/ToolbarDivider/ToolbarDivider.tsx
@@ -1,18 +1,16 @@
 import * as React from 'react';
-import { useToolbarDivider_unstable } from './useToolbarDivider';
-import { renderToolbarDivider_unstable } from './renderToolbarDivider';
 import { useToolbarDividerStyles_unstable } from './useToolbarDividerStyles';
 import type { ToolbarDividerProps } from './ToolbarDivider.types';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
+import { renderDivider_unstable, useDivider_unstable } from '@fluentui/react-divider';
 
 /**
- * ToolbarDivider component - TODO: add more docs
+ * ToolbarDivider component
  */
 export const ToolbarDivider: ForwardRefComponent<ToolbarDividerProps> = React.forwardRef((props, ref) => {
-  const state = useToolbarDivider_unstable(props, ref);
-
+  const state = useDivider_unstable({ vertical: true, ...props }, ref);
   useToolbarDividerStyles_unstable(state);
-  return renderToolbarDivider_unstable(state);
+  return renderDivider_unstable(state);
 });
 
 ToolbarDivider.displayName = 'ToolbarDivider';