+++ packages/react-toolbar/src/components/ToolbarDivider/index.ts
@@ -1,5 +1,3 @@
 export * from './ToolbarDivider';
 export * from './ToolbarDivider.types';
-export * from './renderToolbarDivider';
-export * from './useToolbarDivider';
 export * from './useToolbarDividerStyles';