+++ packages/theme-samples/CHANGELOG.md
@@ -1,8 +1,72 @@
 # Change Log - @fluentui/theme-samples
 
-This log was last generated on Thu, 31 Mar 2022 07:38:07 GMT and should not be manually modified.
+This log was last generated on Mon, 18 Apr 2022 07:38:04 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [8.4.22](https://github.com/microsoft/fluentui/tree/@fluentui/theme-samples_v8.4.22)
+
+Mon, 18 Apr 2022 07:38:04 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/theme-samples_v8.4.21..@fluentui/theme-samples_v8.4.22)
+
+### Patches
+
+- Bump @fluentui/react to v8.65.1 ([PR #22524](https://github.com/microsoft/fluentui/pull/22524) by beachball)
+
+## [8.4.21](https://github.com/microsoft/fluentui/tree/@fluentui/theme-samples_v8.4.21)
+
+Fri, 15 Apr 2022 07:42:48 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/theme-samples_v8.4.20..@fluentui/theme-samples_v8.4.21)
+
+### Patches
+
+- Bump @fluentui/react to v8.65.0 ([PR #22499](https://github.com/microsoft/fluentui/pull/22499) by beachball)
+- Bump @fluentui/scheme-utilities to v8.3.6 ([PR #22499](https://github.com/microsoft/fluentui/pull/22499) by beachball)
+
+## [8.4.20](https://github.com/microsoft/fluentui/tree/@fluentui/theme-samples_v8.4.20)
+
+Thu, 14 Apr 2022 07:38:11 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/theme-samples_v8.4.19..@fluentui/theme-samples_v8.4.20)
+
+### Patches
+
+- Bump @fluentui/react to v8.64.4 ([PR #22501](https://github.com/microsoft/fluentui/pull/22501) by beachball)
+
+## [8.4.19](https://github.com/microsoft/fluentui/tree/@fluentui/theme-samples_v8.4.19)
+
+Tue, 12 Apr 2022 07:39:33 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/theme-samples_v8.4.18..@fluentui/theme-samples_v8.4.19)
+
+### Patches
+
+- Bump @fluentui/react to v8.64.3 ([PR #22439](https://github.com/microsoft/fluentui/pull/22439) by beachball)
+
+## [8.4.18](https://github.com/microsoft/fluentui/tree/@fluentui/theme-samples_v8.4.18)
+
+Mon, 11 Apr 2022 07:40:58 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/theme-samples_v8.4.17..@fluentui/theme-samples_v8.4.18)
+
+### Patches
+
+- Bump @fluentui/react to v8.64.2 ([PR #22446](https://github.com/microsoft/fluentui/pull/22446) by beachball)
+
+## [8.4.17](https://github.com/microsoft/fluentui/tree/@fluentui/theme-samples_v8.4.17)
+
+Wed, 06 Apr 2022 07:34:58 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/theme-samples_v8.4.16..@fluentui/theme-samples_v8.4.17)
+
+### Patches
+
+- Bump @fluentui/react to v8.64.1 ([PR #22266](https://github.com/microsoft/fluentui/pull/22266) by beachball)
+
+## [8.4.16](https://github.com/microsoft/fluentui/tree/@fluentui/theme-samples_v8.4.16)
+
+Mon, 04 Apr 2022 20:01:04 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/theme-samples_v8.4.15..@fluentui/theme-samples_v8.4.16)
+
+### Patches
+
+- Bump @fluentui/react to v8.64.0 ([PR #22295](https://github.com/microsoft/fluentui/pull/22295) by beachball)
 
 ## [8.4.15](https://github.com/microsoft/fluentui/tree/@fluentui/theme-samples_v8.4.15)